requirements.txt
# local package
-e .

# external requirements
click
requests
sqlite-utils
python-dotenv
pytest